requirements.prod.txt
beautifulsoup4===4.12.3
langchain=== 0.2.6
langchain-community===0.2.6
langchain_experimental===0.0.62
playwright===1.41.2
sentence-transformers===3.0.1
tk===0.1.0
pypdf===4.0.1
langchain-openai=== 0.1.13
selenium==4.18.1
pypdf===4.0.1
PyPDF2===3.0.1
uvicorn
fastembed===0.2.1
streamlit===1.39.0
pydantic-settings===2.3.4
cashews===7.4.0
streamlit-feedback===0.1.3